refactor: 404, Home, Category, and Game pages with updated import paths and new UI components
