+++ src/pages/404/index.ts
@@ -1,0 +1,1 @@
+export { Page404 } from './ui/Page404';
