+++ src/pages/404/ui/Page404.tsx
@@ -1,8 +1,9 @@
-import { Box, Button, Typography } from '@mui/material';
 import { Link } from 'react-router-dom';
+import { Box, Button, Typography } from '@mui/material';
+
 import { pathKeys } from '@/shared/lib/react-router';
 
-export function Page404() {
+export const Page404 = () => {
   return (
     <Box>
       <Typography variant="h1" component="h1">
@@ -16,4 +17,4 @@ export function Page404() {
       </Link>
     </Box>
   );
-}
+};
