+++ src/pages/Category/ui/CategoryPage/index.ts
@@ -1,0 +1,1 @@
+export { CategoryPage } from './CategoryPage';
