+++ src/pages/Result/index.ts
