--- src/pages/page-404/index.ts
@@ -1,1 +1,0 @@
-export { Page404 } from './page-404.ui';
