--- src/pages/result/result-page.lib.ts
@@ -1,3 +1,0 @@
-export function isArraysEqual(firstArray: unknown[], secondArray: unknown[]) {
-  return firstArray.toString() === secondArray.toString();
-}
